refactor(catalog): move CatalogModel to catalog and bake in ClaimControlledModel
Relocate CatalogModel from apps/core to apps/catalog/models/_base.py and
add ClaimControlledModel to its bases. "Every catalog entity is
claim-controlled" becomes a structural fact, eliminating
cast(ClaimControlledModel, ...) workarounds in entity_crud.py and
soft_delete.py and letting concrete model base lists drop the redundant
EntityStatusMixin and ClaimControlledModel declarations they were carrying
transitively.

Re-declares EntityStatusMixin's objects: ClassVar[CatalogManager[Self]]
descriptor on CatalogModel so Self rebinds at the catalog level. Without
this, model_cls.objects.active() (where model_cls: type[ModelT:
CatalogModel]) types as CatalogManager[EntityStatusMixin] because mypy
walks the type bound, not the concrete class. EntityStatusMixin keeps its
own declaration so Location (which uses the mixin without CatalogModel)
retains .active() typing.

Pure structural refactor — makemigrations --dry-run reports no changes.
