+++ backend/apps/catalog/models/title.py
@@ -8,8 +8,6 @@ from django.core.validators import MinValueValidator
 from django.db import models
 
 from apps.core.models import (
-    CatalogModel,
-    EntityStatusMixin,
     MarkdownField,
     SluggedModel,
     TimeStampedModel,
@@ -19,7 +17,8 @@ from apps.core.models import (
     status_valid,
 )
 from apps.core.validators import validate_no_mojibake
-from apps.provenance.models import ClaimControlledModel
+
+from ._base import CatalogModel
 
 __all__ = ["Title", "TitleAbbreviation"]
 
@@ -31,8 +30,6 @@ if TYPE_CHECKING:
 
 class Title(
     CatalogModel,
-    EntityStatusMixin,
-    ClaimControlledModel,
     SluggedModel,
     TimeStampedModel,
 ):
